+++ dl_vsix/__init__.py
@@ -1,0 +1,2 @@
+APPNAME = "dl-vsix"
+APPAUTHOR = "sco1"