+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<major>`.`<minor>`.`<patch>`)
 
+## `[v0.2.0]`
+
+### Changed
+* (Internal) Query Gallery API for latest release version prior to download & use for explicit download target
+
+### Added
+* Add extension version to download filename (e.g. `ms-python.python_2025.13.2025082601.vsix`)
+
 ## `[v0.1.0]`
 
 Initial release 🎉